+++ libs/cardano-ledger-core/src/Cardano/Ledger/Genesis.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
@@ -7,6 +8,12 @@ module Cardano.Ledger.Genesis (
   NoGenesis (..),
 ) where
 
+import Cardano.Ledger.Binary (
+  DecCBOR (..),
+  EncCBOR (..),
+  FromCBOR (..),
+  ToCBOR (..),
+ )
 import Cardano.Ledger.Core.Era (Era)
 import Data.Kind (Type)
 
@@ -16,3 +23,13 @@ class Era era => EraGenesis era where
 
 data NoGenesis era = NoGenesis
   deriving (Eq, Show)
+
+instance Era era => ToCBOR (NoGenesis era) where
+  toCBOR _ = toCBOR ()
+
+instance Era era => FromCBOR (NoGenesis era) where
+  fromCBOR = NoGenesis <$ fromCBOR @()
+
+instance Era era => EncCBOR (NoGenesis era)
+
+instance Era era => DecCBOR (NoGenesis era)